+++ docs/dev-dist/sw.js
@@ -82,7 +82,7 @@ define(['./workbox-0b14c527'], (function (workbox) { 'use strict';
     "revision": "3ca0b8505b4bec776b69afdba2768812"
   }, {
     "url": "index.html",
-    "revision": "0.b4s81nckj68"
+    "revision": "0.4bpmoaagme8"
   }], {});
   workbox.cleanupOutdatedCaches();
   workbox.registerRoute(new workbox.NavigationRoute(workbox.createHandlerBoundToURL("index.html"), {