+++ packages/plugin/todo/examples/config.ts
@@ -9,12 +9,17 @@ import { config } from '../src/main'
 const pluginDir = joinPath( getCurrentDir( import.meta.url ), '..'  )
 const wsDir     = joinPath( pluginDir, '..', '..', '..' )
 
-export default defineConfig( config( { all : {
-	input     : [ joinPath( wsDir, '**/*.{js,ts,md}' ) ],
-	output    : joinPath( pluginDir, 'TODO.md' ),
-	inputOpts : {
-		gitignore : true,
-
-		cwd : wsDir,
+// const buildDir  = joinPath( pluginDir, 'build' )
+export default defineConfig( config( {
+	all : {
+		input     : [ joinPath( wsDir, '**/*.{ts,tsx,js,jsx,md}' ) ],
+		// output    : joinPath( buildDir, 'TODO.md' ),
+		inputOpts : { cwd: wsDir },
 	},
-} } ) )
+	docs : {
+		input      : [ joinPath( wsDir, 'docs/**/*.md' ) ],
+		// output    : joinPath( buildDir, 'TODO.md' ),
+		customTags : [ '\\- \\[ \\] \\s*' ],
+		inputOpts  : { cwd: wsDir },
+	},
+} ) )