+++ packages/plugin/todo/src/core/main.ts
@@ -1,4 +1,10 @@
-
+/**
+ * TODO CORE.
+ *
+ * looks for todos in comments in diferent fyle types.
+ * @todo fix for accept "- [ ]" as custom tag without comments
+ */
+import { PluginCore } from '@dovenv/core'
 import {
 	getExtName,
 	getMatch,
@@ -7,20 +13,28 @@ import {
 	relativePath,
 	writeFile,
 	process,
+	getDirName,
+	ensureDir,
 } from '@dovenv/core/utils'
 import {
+	isExtensionSupported,
 	parse,
 	report,
 } from 'leasot'
 
-import type { Config }      from './types'
-import type { TodoComment } from 'leasot/dist/definitions'
+import type { Config } from './types'
 
-export class Todo {
+type TodoComments = Awaited<ReturnType<typeof parse>>
 
-	constructor( public config?: Config ) {}
+export class Todo extends PluginCore {
 
-	async get( key?: string[] ) {
+	constructor( public config?: Config ) {
+
+		super()
+
+	}
+
+	async get( key?: string[] ): Promise<TodoComments | undefined> {
 
 		const conf = this.config
 		if ( !conf ) {
@@ -30,23 +44,43 @@ export class Todo {
 
 		}
 
-		const availableKeys      = Object.keys( conf )
-		const currKeys           = ( !key ) ? availableKeys : getMatch( availableKeys, key  )
-		const res: TodoComment[] = []
+		const defaultType            = '.md'
+		const availableKeys          = Object.keys( conf )
+		const currKeys               = ( !key ) ? availableKeys : getMatch( availableKeys, key  )
+		const resTotal: TodoComments = []
+
 		for ( const key of currKeys ) {
 
-			console.info( `TODOs for [${key}]` )
+			const res: TodoComments = []
+			console.log( '\n' + this.style.get.section( `TODOs for ${this.style.get.badge( key )} key` ) )
+
 			const opts  = conf[key]
-			const paths = await getPaths( opts.input, opts.inputOpts )
+			const paths = await getPaths( opts.input, {
+				gitignore : true,
+				onlyFiles : true,
+				dot       : true,
+				...opts.inputOpts,
+			} )
+
+			console.debug( { paths } )
 
 			for ( const path of paths ) {
 
-				const content  = await readFile( path, 'utf-8' )
-				const filetype = getExtName( path )
+				const content = await readFile( path, 'utf-8' )
+				let filetype  = getExtName( path )
 
-				const todos = parse( content, {
+				// if ( !filetype ) console.warn( 'path has no file type', path )
+
+				const fileTypeSupported = isExtensionSupported( filetype )
+				if ( !fileTypeSupported ) {
+
+					console.warn( 'path has invalid file type', path )
+					filetype = defaultType
+
+				}
+				const todos = await parse( content, {
 					filename   : relativePath( process.cwd(), path ),
-					extension  : filetype,
+					extension  : filetype || defaultType,
 					customTags : ( opts.customTags
 						? opts.customTags
 						: [
@@ -54,7 +88,6 @@ export class Todo {
 							'@todo',
 							'@fixme',
 							'- [ ]',
-							'- [x]',
 						] ),
 				} )
 
@@ -66,24 +99,44 @@ export class Todo {
 
 			if ( res.length > 0 && opts.output ) {
 
-				const output = report( res, 'markdown', { padding: 2 } )
+				const output = await report( res, 'markdown', { padding: 2 } )
+				const outDir = getDirName( opts.output )
+				await ensureDir( outDir )
 				await writeFile( opts.output, output, 'utf-8' )
 
 			}
 
+			if ( res.length > 0 ) {
+
+				const output = await report( res, 'table', { padding: 2 } )
+
+				if ( output && output.length > 0 )
+					console.log( output )
+
+			}
+
+			if ( res.length === 0 ) console.log(  '\n' + this.style.get.error( 'No TODOs found' ) )
+			else resTotal.push( ...res )
+
 		}
-		if ( res.length === 0 ) console.log( 'No TODOs found' )
-		return res
+
+		return resTotal
 
 	}
 
 	async run( key?: string[] ) {
 
-		const todos = await this.get( key )
-		if ( !todos ) return
-		const output = report( todos, 'table', { padding: 2 } )
-		if ( output || output.length > 0 )
-			console.log( output )
+		try {
+
+			await this.get( key )
+
+		}
+		catch ( e ) {
+
+			console.log()
+			console.log( this.style.get.error( e instanceof Error ? e.message : e ) )
+
+		}
 
 	}