+++ packages/plugin/todo/src/core/types.ts
@@ -3,15 +3,25 @@ import type { getPaths } from '@dovenv/core/utils'
 export type Config = {
 	/** Id  of the TODO */
 	[key in string]: {
-		/** Input pattern for get TODOs */
+		/**
+		 Input pattern for get TODOs
+		 */
 		input       : string[]
 		/**
 		 * Output for add all TODOs in markdown format.
 		 * Override if exists content.
-		 * @example './TODO.md'
+		 * @example undefined
 		 */
 		output?     : string
-		/** Options for get TODOs */
+		/**
+		 * Options for get TODOs
+		 * @default
+		 * {
+		 *   gitignore : true,
+		 *   onlyFiles : true,
+		 *   dot       : true,
+		 * }
+		 */
 		inputOpts?  : Parameters<typeof getPaths>[1]
 		/**
 		 * custom tags to get TODOs