+++ packages/plugin/examples/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @dovenv/examples
 
+## 0.0.16
+
+### Patch Changes
+
+- Fix todo and docs plugins
+
+- Updated dependencies []:
+  - @dovenv/core@0.0.16
+
 ## 0.0.15
 
 ### Patch Changes