+++ packages/plugin/lint/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @dovenv/lint
 
+## 0.0.16
+
+### Patch Changes
+
+- Fix todo and docs plugins
+
+- Updated dependencies []:
+  - @dovenv/eslint-config@0.0.15
+  - @dovenv/stylelint-config@0.0.15
+  - @dovenv/core@0.0.16
+
 ## 0.0.15
 
 ### Patch Changes